Bugfix: negative counts when using front crossings.
When we counted front crossings, it was possible to end up with negative
or otherwise incorrect counts for a couple reasons:
- improperly checking whether an object crossed a boundary or not
- continue checking tentative collisions when we should stop

Note 1: The counts actually involving the crossings were correct. Their
presence however could make other counts incorrect.

Note 2: there are very likely other problems in the cases where we call
count_tentative_collisions. This function is part of the recent
refactoring of diffuse_3D. The problem with count_tentative_collisions
is that it replaced a lot of code which was *nearly* but not exactly
identical. It is these slight differences that are accounting for these
recent bug fixes. Expect more fixes to follow as these problems are
identified.