Bugfix: previous PB being over-written.
A couple things need to be explained to understand this bug. First, if a
surface molecule can't be placed (e.g. if the tile is occupied), then we
try to place it again (pick a new displacement, 2d raytrace,
bookkeeping, etc). Second, to check if a molecule has entered a new
periodic box, we always keep track of the previous box it was in. The
problem was that the previous box that we were in was being over-written
when we tried to place the molecule again.