optimizing how hashmap is handled
this commit introduces some preprocessing steps for calculating a hash based
on the graph pattern string for every reactant in a reaction.
Bimolecular reaction hashes are then calculated by adding the individual hashes
of every reactant, while unimolecular hashes are directly the hash of its reactant.
This avoids having to create a string key every step which drastically increases performance.