modified to subvolume hash species list to use the graph label for its key
mcell-nfsim particles are all based on the same base mcell-species. a proxy or sorts
this was causing that all the particles associated with a given subvolume were assigned
to the same species list. this change makes it so that if a given species has the EXTERNAL
tag on, it will instead organize the associated subvolume hash table on a graph-label basis.

We are  taking advantage of this in the reaction_trigger_preliminary_nfsim function, a function
that checks a given particles against all lists in the same subvolume. This cuts on processing time
by avoiding unnecessary comparisons against particles we can't react with anyway.