Produce error if using PBCs with checkpoints.
Checkpointing would definitely not work in the case of non-traditional
periodic boundary conditions, since we are not (yet) storing the
molecule's current periodic box in a checkpoint file. I think the
traditional form would probably work, but it needs to be tested first,
so I'm just disabling it for now.