SMs still being released outside PBs.
This is an update to commit a90afd28. There were a couple problems with
the previous solution. First of all, it didn't account for the case
where it's hard to find a spot to place the molecule (i.e. when the
"seek_cost >= pick_cost"). Secondly, we were grabbing one random
molecule position in release_onto_regions and another different one in
place_single_molecule. As such, I'm surprised it worked as well as it
did. :/ Needless to say, all of this should probably be tested more
thoroughly.