Bugfix: surface molecules escaping PBs.
The imprecision involved in converting back and forth between UV and XYZ
coordinates led to molecules escaping the periodic boundaries. More
specifically, when an SM hit the periodic boundary, we would convert
that XYZ position to a UV position and treat it as the molecule's new UV
position. Later, we would convert that UV position back to an XYZ
position. A molecule which was previously at the periodic boundary would
suddenly be slightly outside the boundary because of this imprecision in
conversion.