Make it an error to release a VM outside of PB.
Make releasing a volume molecule outside of a periodic boundary an
actual error message.

Unfortunately, it's not as straightforward doing this for surface
molecules, since you essentially need to have meshes that pass
through/beyond the periodic boundariesi. That means we might attempt to
place SMs outside of the boundaries if we're just picking a random tile
on an object. If this happens, we don't give up; we just re-try the
placement.