Bugfix: count SMs wrong when release not at origin
When releasing surface molecules in a periodic box other than the
"origin" box ([0,0,0]), the molecule counts would be incorrect, because
the periodic box was being set after setting up the count. This was
already partially working, but it didn't work in certain instances (e.g.
list based releases). This is very similar to commit c38e65d6, except
for surface molecules instead of volume molecules.