Multiple SMs at same grid loc, diff PBs.
You should be able to have multiple surface molecules at the same grid
location so long as they are in different periodic boxes. This was
partially working previously in a hackish sort of way. This commit
considerably improves the situation, but it meant rewiring a lot of
MCell guts in a way that might not be terribly efficient. As such, this
solution should be tested more extensively. Additionally, a few of the
regular nutmeg tests involving surface molecules are now broken meaning
that additional MCell fixes will be required.