Have Travis use Coverity.
When a build is submitted to Travis, it should (if this works) also do a
static analysis scan on Coverity.

In the future, we may want to consider only doing this on a special
branch (e.g. named coverity_scan or something like that) since you only
get ~12 builds/week for a project of our size on Coverity's service. For
now, it should be fine since most of our work goes into feature
branches.