Fix memory leaks & possible null pointer problems.
Fixing more memory leaks. Also, there were several places where we were
dereferencing a pointer and then checking after the fact if it was a
null pointer (e.g. var->x = x; if (var == NULL) { do_stuff }) or where
there was a possibility that a pointer was null and then it would be
dereferenced.