First step at getting PBCs to work with DGs.
This is the first step at getting periodic boundary conditions to work
with dynamic geometries. So far, it's only been tested with surface
molecules and traditional PBCs. There are currently some cases where it
will fail because it tries to snap a molecule to a spot outside of the
periodic box after changing the meshes for a dynamic geometry event.