Fix more problems with DGs and nontrad PBCs.
This seems to fix the remaining problem with counting volume molecules
using dynamic geometries and non-traditional periodic boundary
conditions.

Unfortunately, it uses count_region_from_scratch, so it's almost
certainly a bit slower. Having said that, I don't know if there's anyway
around it for the time being. The problem, as far as I can tell, has to
do with the fact that when using dynamic geometries with PBCs you might
have meshes that cut across the boundaries of the periodic box. If you
do, then you aren't getting the proper count statements for a molecule
as it crosses between PBs while still inside a single reflective mesh
(that cuts through the PB). Normally, you have to cross a transparent
mesh boundary to get that update, but that's never going to happen in
this case.

                 PB
            |-----------|
            |           |
        |---|-----------|--|
        |   |           |  |mesh cutting through PB
        |---|-----------|--|
            |           |
            |-----------|