mcell_modify_rate_constant works with unimol rxns.
When rate constants are changed by mcell_modify_rate_constant, the
lifetimes of molecules involved in unimolecular reactions are recomputed
now. This involved writing a new function for sched_util.c that just
"peaks" at things in the scheduler without advancing the scheduler or
updating the current_count. Although everything seems to be working,
additional testing should still be done.