Use Python 3. Bugfix with mcell_run_iteration.
pyMCell now uses Python 3 instead of Python 2. This makes sense since
that's what we are using with CellBlender. Also, we're currently using
the Python 3 version of pyNEURON.

There was a strange bug with mcell_run_iteration where the typemap for
mcell_run_iteration wasn't working. The typemap was replaced by using
the special swig keyword INPUT.