Update cmake version reqs. Relax python reqs.
Getting pyMCell to build with older versions of cmake (like 2.8) seems
to be really problematic. I've updated the cmake requirement to 3.4.
That might be unnecessarily high, but I haven't had a chance to test it
on versions between 2.8 and 3.4.

I've also relaxed the Python requirements from Python 3.5 to any version
of Python 3.