Added some of the earlier C++ molecule/reaction API from 2013
The C++ classes from the 2013 API were added mostly to include them as
reference material for one implementation option.

This version uses operator overloading of +, >, and == for the reaction
syntax of +, ->, and <->. So the following legal C++ code would
represent an MCell reaction definition:

   new_reaction ( a + b  >  c + d )

where a, b, c, and d represent species, and a+b react to become c+d.

This implementation has some nice features (such as compile time
checking), but it is likely that the more familiar string expressions
will become the final implementation.