Added new "CLAMP_FLUX" boundary condition for volume molecules
This is the first commit of new code to implement a CLAMP_FLUX option to
surface classes.  This the new code shares common code with the
CLAMP_CONCENTRATION option.  Consequently I renamed many common functions
and data structures to be more general regarding "clamps".

CLAMP_CONCENTRATION implements a Dirichlet (i.e. constant value) boundary
CLAMP_FLUX implements a Neumann (i.e. constant flux) boundary

In both cases we implement the boundary condition using a constant flux due to
a constant concentration coming from an imaginary source on the other side of
the boundary.  In the case of CLAMP_CONCENTRATION all particles hitting the
surface from the clamped side of the surface are absorbed, while in the case of
CLAMP_FLUX all particles hitting the surface from the clamped side are
reflected.