Working builds of pymcell and mcell executable.
This is less than ideal, but at least test_api.py and the DG tests in
nutmeg work correctly. There's a problem with enabling all the normal
functionality of dynamic meshes to pymcell because DG uses the parser
directly (parse_input and mdlparse_file), which doesn't seem to be
possible to do with swig (or at least I haven't figured out a way to do
it). Besides, this isn't really how you'd want to run a simulation with
pymcell.