Trying to generalize last commit.
In the last commit, I managed to get DGs to work with pymcell, but the
geometry swap was hardcoded in a hackish way. This generalizes the
process a bit, although it seems to be a bit broken currently. It
probably has to do with memory being corrupted.

To use this new process, a user first calls mcell_destroy_everything,
then adds new geometry, and finally calls mcell_reinitialize.