Molec counts cause geom changes. Fix memory probs.
There's still a way to go, but this is (hopefully) a step in the right
direction. You can actually have molecule counts/concentrations dictate
geometry changes. There's less being passed between the C code and
python during the actual dynamic geometry process, which ameliorated a
lot of the memory problems. Right now, you can only pass in one object,
so this is more like a proof of concept than anything.