Allow multiple objs to be updated via DG-pymcell.
Previously, you could only update a single object and the region name
was hard coded. The code in this commit could still use some cleanup and
additional testing.

When you run test_api.py, it creates a simulation with a torus and a
cube. Molecules capable of undergoing a bimolecular reaction
(vm1+vm2->vm3) are released in both objects. When the concentration of
the product (vm3) exceeds some arbitrary threshold, the location of the
torus and cube are updated for every subsequent iteration.