Reverse vertex and face list.
This is a simple fix to reverse the vertex and face lists that are
passed to create_polygon_object (which is more consistent with how it's
done in the parser). This change has been made at the Python level but
should probably be re-done at the C level (e.g.
mcell_add_to_connection_list). I'm not sure if this explains all the
weirdness that Oliver experienced, so it would be worth looking into
this matter further.