Bugfix: rxns with non-diffusing molecules.
When trying to change the rate constant of non-diffusing molecules, it
wouldn't actually change because we were only checking for molecules
coming up next in the scheduler. This works for diffusing molecules, but
not non-diffusing molecules (unless their lifetime coincidentally
happens to be up).

The current solution is inefficient and needs to be optimized. I want to
make sure it works for all test cases before worrying about that though.