Only recompute lifetimes if we really need to.
The solution is still inefficient, but now it's only triggered if it
really needs to be done (i.e. in cases where you have a unimolecular
reaction x->y or when you have a surface molecule reacting with a
surface sm@sc->z). As before, this still really needs a lot more
testing.