Modified to get arguments to print (used "bare" instead of "wrapped").
This version simply prints the current location (getcwd) and the command
line parameters before anything else. This is helpful when running array
jobs that sweep a parameter space.

This code was attempted after the "mcell_print_version" call using the
standard MCell log functions, but failed to produce output. This version
worked.