Bugfix: all (vol) mol surf class + rxn.
If you have a surface class that is transparent/absorptive to all
(volume) molecules, you shouldn't be able to use it in a reaction. This
worked in some instance, but not others. This was found out by a user in
our forums here:

http://mmbios.org/mcell-cellblender-forum/visualization/104-modelling-objects-how-to-make-plane-vertical

Note: In this section of the code, the logic is extremely convoluted and
hard to follow. It would be useful to refactor this at some point. Oh,
and add more tests!!!

P.S. Also made small grammar tweaks to the error messages.