Added an MCell logo file in the doxygen/html file and ignored others.
Adding a file in a directory will cause git status to list all other
files in that directory as untracked. So adding the logo image from the
doxygen/html folder will list all of the generated html as untracked. To
avoid this, the docs/doxygen folder was added to the .gitignore file.