Initial version of BNGL viz output - Needs lots of cleanup / freeing.
This version implements the visualization approach discussed in our most
recent MCell/CellBlender meeting. In particular, it draws each of the
molecules in a complex as separate glyphs. When a complex forms, the
previously separate glyphs will be drawn at the same location.

This version has been tested with the fceri model and the complex of
Lyn.Rec was observed to form on the surface of the cube.

This version also continues to draw the two proxy molecules (volume
and surface). While they do not show up in the Define Molecules panel,
they volume and surface proxies can be manipulated through Blender's
outliner to show, hide, or resize. This provides a nice check on the
drawing of the actual molecule types.

Finally, this is a very preliminary version and does NOT free up any
of the memory allocated to build the separate molecules. That should
be the next step.