Removed print statements used during debugging. This version works.
This version appears to work properly, and it produces viz files with
both proxy molecules and individual molecule parts that make up
complexes. It has been tested with CellBlender and the visualization
is reasonably good - especially with the right choice of glyphs and
relative molecule sizes. This should be sufficient for the workshop.