Add unique molecule id to molecule trigger output
We were already tracking the molecule IDs, but it was currently only
made available in the viz data. This adds it to the final column of the
trigger output involving molecules (doesn't modify triggers for
reactions or hits/crossings).

The motivation for this is that it will allow single particle tracking.
Between the IDs in the trigger output and the viz data, individuals
should be able to track a single particle throughout the entire
simulation as it undergoes reactions using some simple scripts (to be
written).

Eventually, once we figure out how this will be used, we could include
some of this analysis functionality in CellBlender.