Add a MANIFEST.in for creating python dist.
When you run "python setup.py sdist", python will look for a MANIFEST
file to figure out what files to include in the tarball it builds. If it
doesn't exist, it will create one based off of the rules in the
MANIFEST.in file.

The motivation for making these tarballs is that this is the format they
need to be in to be uploaded to PyPI and thus available to the world via
a "pip3 install pymcell".