Simplify and clean up API.
This is the beginning of an effort to clean up the pymcell API. It is
still too close to the underlying c API, which is clunky and hard to
use. We can layer some simple abstractions on top of it to hide a lot of
the ugliness. Ultimately, we'll want to clean up the c api itself, but
by tackling the python api first, we can prototype the changes much more
quickly.

Right now, this simplification of the API is partially functional and
allows you to define molecules, reactions, geometry, and run the
simulation. An example of its use is in test_new_api.py.