Automatic simulation cleanup.
When a user would forget to run all the iterations specified and the
script would exit, it would cause the terminal to hang. This commit
tries to address the problem by using __del__ to call end_sim. We should
probably solve this problem with a context manager instead, since
__del__ can be so problematic. In this case though, it seems to prevent
the hanging problem (at least with CPython).