Resolve nfsim library issues for setup.py.
Even though the normal cmake build has been working fine in this branch
for a while, setup.py has been kind of broken due to nfsim library
issues. It's important to have setup.py working, since it's required to
make a build for PyPI (thus allowing users to do a "pip install
pymcell"). Note: the current solution feels pretty hackish, but I'm not
sure what the best practices are for something like this.