Use an rpath for nfsim libraries.
This will allow the library files to be moved around relative to the
binary, which is important for bundling with CellBlender.

Also, reverted back to SWIG_ADD_MODULE, since the version of CMake that
you can get via the repos in Ubuntu 16.04 doesn't support
SWIG_ADD_LIBRARY.

Lastly, I added the version.h dependency for pymcell.