Update dependencies in README. Fix travis.yml.
From the list of dependencies, we were missing boost, swig, and other
stuff needed to build libNFsim and pyMCell.

Also, I clarified a few things relating to pyMCell.

Lastly, travis.yml has been broken for a little while partially because
of nutmeg and also because of the libnfsim and pymcell changes. This
should fix some of the problems, but I need the feedback from travis
itself to be certain.