Clean up pyMCell examples.
These were just thrown into src, which was okay when there was only one
example, but now we have over a dozen different files, so we should
rethink how this is organized and structured.

Note: we use the word test a lot with these files, but we should maybe
change that so it doesn't get confused with the unit tests.