Update files needed by appveyor/windows builds.
These are files that need to be pre-generated on a linux machine in
order to build MCell on Windows. It's mainly flex and bison stuff. Now,
that the Windows build system is being revisited, we should find a more
robust solution to this problem, possibly using flex and bison directly
on Windows.