CMake working on Windows.
If you have winflexbison installed, this should hopefully work fine.

There are still some things that need to be cleaned up. For instance, we
should stop just copying over a pre-existing copy of version.h and
create it with a Python script or something along those lines. Also, we
should better document all the Windows dependencies in the README. Right
now, the Windows build process is buried in the github wiki.

I removed the sed bit, which was in there for the autotools build
system, but that method is pretty deprecated at this point. I'm not even
sure if it works anymore.