Use the correct config.h file for Windows
Apparently, we had the config-nix.h file in the appveyor dir. Whoops. It
shouldn't matter for the cmake build, since it's getting copied over
from src/config-win.h. This is just to be extra safe. Eventually, we
shouldn't have any files in the appveyor dir and just copy/generate
them directly.