Converted previous viz code to use symbol tables, added dump_symtab.
The previous version stored a list of molecules that had been seen in a
simple linked list. This version uses a sym_table_head structure to save
the same information.

The graph_pattern_table holds entries of all the unique molecules found
during a simulation. They are keyed by name, and the data is currently
just a NULL stored as a VOID_PTR type. Eventually this will be replaced
with a C structure to hold the spatial structure.