Stored the graph parts in an array for further processing.
The previous version simply printed the graph parts for each graph
string (NAUTY string). This version allocates an array of strings and
fills it with each graph part for further processing and frees it
afterward. The "further processing" in this case is just printing, but
future versions will build a tree structure representing the structured
molecule.