Decoded the molecule/component names and connections as an array.
The graph pattern is a long string (broken here by commas):

  c:SH2~NO_STATE!12,c:SH2~NO_STATE!11!2,c:SH2~NO_STATE!13!1,
  c:SH2~NO_STATE!14!9,c:U~NO_STATE!11!5,c:U~NO_STATE!12!4,
  c:U~NO_STATE!13!7,c:U~NO_STATE!14!6,c:a~NO_STATE!15,c:b~pY!15!3,
  c:g~Y!15,m:Lyn@PM!1!4,m:Lyn@PM!0!5,m:Lyn@PM!2!6,m:Lyn@PM!3!7,m:Rec@PM!8!9!10,

This commit parses it into an array of parts connected by references:

  c[0]: SH2   [12]
  c[1]: SH2   [11]   [2]
  c[2]: SH2   [13]   [1]
  c[3]: SH2   [14]   [9]
  c[4]: U   [11]   [5]
  c[5]: U   [12]   [4]
  c[6]: U   [13]   [7]
  c[7]: U   [14]   [6]
  c[8]: a   [15]
  c[9]: b   [15]   [3]
  c[10]: g   [15]
  m[11]: Lyn@PM   [1]   [4]
  m[12]: Lyn@PM   [0]   [5]
  m[13]: Lyn@PM   [2]   [6]
  m[14]: Lyn@PM   [3]   [7]
  m[15]: Rec@PM   [8]   [9]   [10]

The numbers in brackets [] are either addresses (on left) or references
to addresses (on right). Since the entire graph is connected, any
starting point will eventually lead to all others.

Starting, for example, at c[6] (component U), it's connected to:
  [13] and [7]
   [13] is Lyn which is connected to [2] (SH2) and [6] (previous U)
   [7] is U which is connected to [14] (Lyn) and [6] (previous U)

This can be continued through the entire complex. This version of the
code uses an "external_molcomp_loc_struct" to store both molecule and
component locations. The only difference is a flag ("is_mol").
Otherwise, they each (molecules and components) have a location, name,
graph_string, and peers. The peers are currently indexes in the array
containing all of the molecules and components (as shown above). This
should be a somewhat convenient structure for filling in the actual
locations of each molecule and component based on the spatial structure
defined by CellBlender. Once those values are filled in, then either the
molecules or components or both can be included in the viz data file.