Display the molecules from the new structure with random locations.
This version properly displays structured molecules according to the
coordinates and components in the "external_molcomp_loc" list built
for each species found in the simulation.

Each molecule or component is represented as a "molcomp" structure:

   bool is_mol          // is this item a molecule or component
   double x, y, z       // relative location of this item
   char *name           // short name of this item
   char *graph_string   // full graph string of this item
   int num_peers        // number of connected items
   int *peers           // index numbers of connected items

A structured molecule is represented as a list of these items (actually
stored in an array with a count). The index numbers of the peers are the
index numbers in the list of the entire structured molecule. The peers
connect in both directions (molecules to components, components to
molecules, and components to components in bonds).

This structure would be sufficient to show both molecules and
components, but only the molecules are stored in the viz file because
the components don't have a molecule name for selecting a glyph. We
could use a "generic" molecule (like volume_proxy) so that all
components are optionally displayed using the same glyph.

The current version doesn't yet have a way to provide coordinates for
these structured molecules from CellBlender (or MDL), so the values are
chosen at random for each new structured molecule. But whatever random
coordinates are chosen are stored and used for all molecules of that
type (those with identical NAUTY strings). In other words, the code to
generate the coordinates based on some kind of assembly patterns has yet
to be done. But all of the infrastructure is in place to support that
code when it becomes available.

Note that this version does a fair amount of dynamic allocation, so it
would be wise to check for memory leaks.