Worked out most of a 2D molecule positioning approach (stub), but ...
This version attempts to lay out molecules automatically in 2D by
assigning components to positions on a circle around the molecule.
Components are spaced evenly starting from angle=0.

This version compiles and runs, but the molecules appear to be folded
back about the components. The result of the folding is that the folded
molecules are at the same locations and appear to be just one molecule.

Fixing this may require tracking the two ends of each bond separately.

Note that this 2D positioning approach is an experiment to begin to
understand the problems we'll find with the full 3D positioning. The
"folding" found here is one of the problems to be addressed.