Worked on a new way to set the molcomp positions in 2D.
This was an attempt to automatically build molcomp positions in 2D.

To help debug this, all components were dumped in the section of code
commented as "/* Force output of components */" (set to 1). However,
this created many more "component" molecules than expected, so the logic
has been again disabled (set to 0).