Added Java code from SpatialMols2D.java (doesn't work properly yet).
This version includes the Java code for building spatial 2D molecules,
but it does not work yet. At this point, the "acos" function appears to
be returning a "nan".

Also note that this version commented out the following warning:

   "Warning! You have a binding rxn __ that allows a moleculeType to
             bind another of the same type."

This was happening all the time which caused the output to be difficult
to read. Note that while this change was made in this commit, the file
containing the change (../nfsim/src/NFcore/reactionClass.cpp) was not
part of the GIT repository. It may be a submodule. So while it did
compile and work, the source is not directly part of this commit.