Python parsing of component loc and rot in MDLR.
This complements the work that Tom did in commit
84759b69435a6ca8c367f89da0567d252df510c3 and other subsequent commits.

The mdlr2mdl.py allows users to create MDLs from an MDLR file. With this
commit, the script will also create a DEFINE_MOLECULE_STRUCTURES block
in the output molecule MDL by parsing the component locations and
rotations listed in the #DEFINE_MOLECULES block of an MDLR file.

Essentially, you go from something like this (in MDLR):

  #DEFINE_MOLECULES
  {
    Lig(
      l{loc=[xl1, yl1, zl1],rot=[xr1, yr1, zr1, angle1]},
      l{loc=[xl2, yl2, zl2],rot=[xr2, yr2, zr2, angle2]}
    )
  }

to this (in MDL):

  DEFINE_MOLECULE_STRUCTURES {
    Lig(l{loc=[xl1, yl1, zl1], rot=[xr1, yr1, zr1, angle1]}, l{loc=[xl2, yl2, zl2], rot=[xr2, yr2, zr2, angle2]})
    {
            DIFFUSION_CONSTANT_3D = 8.51E-7
    }
  }

TODO:
  - Component states (e.g. ~p) are currently ignored/dropped
  - The formatting of the output could be prettied up
  - Improve error handling
  - Clean up code to be more consistent w/ the rest of Jose's code
  - Right now, loc and rot is optional, but I'm not sure if it should be
  - MCell doesn't seem to be able to parse this yet. I'm not sure if
    it's a bug in my implementation or Tom's. Currently getting this
    error w/ fceri model: Error: BNGL Molecule component already defined: Lig(l)