Added forced file flush for testing, also print "=#= New Graph Pattern"
There have been some suspected race conditions that cause MCellR to
appear to fail at random times. One concern was possible file I/O when
generating the MDl, so this change forced a flush. It did not make any
difference.

This change also re-introduces a little more debug output to print the
graph strings when new molecules are found. This appears to add a small
fraction to the existing output, and it's useful for getting at the
actual graph strings that MCell is trying to display.