Added viz_output code to try to dump the mol_ss tables.
This version contains a new function named:

  "set_component_positions_by_table"

which will replace the "set_components_positions_2D" stub.

At this point, the function still makes the same 2D assignments, but it
attempts to dump the mol_ss table contained in the "world" structure.

As of this version, it's printing values that don't look correct. The
reason is still under investigation.