Added dump option to mcell to allow additional (non-debug) output.
The viz_output code can produce text output useful for importing into
other programs. For example, it can print the complex molecule data in
a manner that can be imported into other visualization software. This
output isn't really "debug" since it's not intended to "debug" MCell.
Instead, it's data that would not otherwise be needed in a normal
simulation.

It's not clear at this time whether the dump level should be used to
specify individual items to be dumped (maybe as a bit mask) or if it
should be a monotonically increasing value where higher values simply
increase the amount of output. It's possible that both approaches might
be used (where higher is generally more, but certain high output data
might only be printed when a certain bit is also set).