Added full support for -d (dump) option in MCell
The "-d #" command line parameter is an optional argument specifying a
"dump level" for MCell. The "dump_level" can be any number, and it's up
to any section of code to determine what to do with any particular value.

In this commit, the viz_output code will print additional information as
follows:

  dump_level >= 10:  Prints all new Graph Patterns as they are detected
  dump_level >= 20:  Prints the molcomp_loc_array (with molecule structure)
  dump_level >= 50:  Prints message at top and bottom of viz output loop

If the -d option is not set, it defaults to 0, and all output should be
the same as it was before this change.