Spatial Molecules: Added comments and renamed a variable for clarity.
The "part_num" variable was used in two different places for different
reasons. In the first use it was an integer really counting up the
number of parts. In the second use it was also an integer but stepping
through the part numbers. The first was changed to "num_parts" to reduce
confusion.