Rotation keys implemented with a forced bond angle of 0.2+0.2 radians.
This commit implements the rotation key processing using a fixed
rotation angle of 0.2 + 0.2 radians. These are expressed as angles from
the perspectives of the two binding molecules, and are ultimately added
together to form a final rotation of 0.4 radians.

Future commits will reset these default rotation angles to 0 and get
them from some other source (MDL processing or a command line option).

The angles were left at 0.2 + 0.2 in this commit because that's the
value used to produce the spiral test results.