Added a command line option of -b #  (or --bond_angle #) for bond angles
The previous version used a "hard coded" bond angle of 0.4 (0.2+0.2).
This version accepts an optional command line parameter (-b) which takes
a bond angle as an argument. The default is 0, and angles can be either
positive or negative. The bond angle is "logically" split between the
two binding components, but the angles are summed eventually anyway, so
the result is that the specified angle will be the angle between the two
reference planes.