Added code to create a spatial subdirectory in viz_data/seed_#
This commit begins to create an alternate viz file format in addition to
the currently existing format. The current code is only triggered with
the "-z opts" command line option where "opts" is non-zero. The "opts"
parameter is intended to provide a way to turn different viz modes on
and off relatively easily during development.

The original plan was to add more files to the viz_data/seed_# directory
to represent the spatially structured molecule data. However, the
current viz code assumes that ALL files found in that directory are viz
files. So creating the viz files there caused CellBlender to assume that
there were twice the number of iterations and half of them produced
errors because the file format wasn't correct.

The next plan was to create a "spatial" subdirectory in the
viz_data/seed_# folder, but the path available already had the "Scene"
part stuck to it. Rather than pull that apart, this version ends up
creating a subdirectory named "Scene_spatial" in the viz_data/seed_#
directory. That's where it creates the "cellss.##.dat" files along with
the "dictionary" of molecule types named "cellssd.dat".