Added ability to show components as "mol_component" molecules.
This commit adds code to create special "mol_component" molecules in the
viz output files whenever the "-d" option is greater than 0.

This code had been mostly added in earlier commits, but a logic error
caused it to enter an infinite loop. This commit fixes the logic error
and attaches this new functionality to the "-d" (dump) option. It might
be better attached to the "-z" option as one of the display modes.

The current version just creates one single "component" molecule named
"mol_component". Because this is not a normal CellBlender molecule, it's
display options must be set using Blender's normal mechanisms (not in
CellBlender's Molecule display options). This does, however, easily
allow changing size, color, and visibility.