Shifted the new file format option to the second LSN (nibble).
The previous version had used bit 0 in the least significant hex digit
(nibble) to specify the new output format which is just getting started.
The addition of the component visibility modes seemed to be a better use
for those lowest bits, so the new file format option is being moved up
to bit 0 in the second least significant hex digit (nibble). This makes
it independent of the component visibility modes.

Note that the "-z" command line option is intended to support rapid
development and testing of many different visualization modes. Using
this approach avoids having to define CellBlender controls and having to
change the MCell MDL parser for new features. At the same time, it's a
bit awkward to specify any number of options in a single integer. But it
turns out that command line integers can be specified in hex using the
familiar C syntax (0x2b3). This is a more convenient way to turn
specific bits (features) on and off independently using the "Command
Options" field in CellBlender's Advanced Options Run panel. Since hex
works well, it's anticipated that options for different types of display
features might be best grouped 4 bits at a time. This allows relatively
easy separation of features in the interface.