Use viz_bngl path and Scene.bnglviz, commented unused functions.
The BNGL viz files will be in a subdirectory:

  ./viz_data/seed_###/viz_bngl/

The file names will be: Scene.bnglviz.###.dat

This version creates those files, but also creates the cellssd.dat file
which was intended as a spatially structured dictionary to hold molecule
structures (and other data) indexed by integers. This will likely be
removed for simplicity.