Removed the space_struct_dict_file variable and references.
The space_struct_dict_file was intended to be a dictionary of spatially
structured molecules referenced by an index number. This would eliminate
the need to compute the structure for each molecule on the fly.

However, now that the new approach is to allow CellBlender to construct
the spatial structure, this dictionary isn't needed (at least isn't
needed to be built by MCell).