Export text versions of the viz molecule header and viz molecule list.
This version creates text files named viz_bngl/Scene.bnglviz.###.dat
containing a header describing each molecule's geometry and a list of
molecule instances with references to header entries. The header entries
are both Nauty string format and "MolComp" format. The MolComp format
has coordinates calculated by MCell but does not have state information.
The Nauty string format has states but no coordinates. Obviously, it
would be good to come up with a convenient format encoding both.

This version generates the header from all of the molecule types that
have been created up to that iteration. Depending on the simulation,
this may be a good or not-so-good strategy. A future version might only
include the types actually used in that particular iteration file.

Right now the files are in text format for easier development. In the
future they would likely be either a binary format or possibly a format
that's easily imported into Python (maybe source code, pickle, or JSON).