Properly handle non-spatial binding between non-spatial molecules.
The earliest versions of MCellR mol-viz simply created all molecules in
a complex at the same location. This was typically visualized by using
different glyphs (torus and sphere, for example) to show the parts.

More recently, the "stick and ball" model was added to allow the spatial
structure to be built and displayed. However, non-spatial molecules
(BNGL molecules with no coordinates) were not handled by that code, and
there was no provision for displaying them properly.

This commit determines that any molecule with any components at the origin
(coincident with the molecule itself) should be considered as non-spatial.
Any non-spatial molecules in a complex currently force the entire complex
to be considered (and displayed) as non-spatial.

This commit fixes the improper display of the "FCERI" non-spatial model
in the examples panel.