Added JSON List Viz Output format (mcell -z 0x23 -d 10)
This version can output a JSON/Python list of lists (no dictionaries,
and no booleans) as one of its formats. The format is:

[
  # File format number
  ffn
  # List of molecule complex templates
  [
    [
      [ 'm', 'A', [0,0,0], [1,2,3], []        ],
      [ 'c', 'x', [0,0,0], [0,5],   ["n"]     ],
      [ 'c', 'y', [0,0,0], [0],     ["m","n"] ],
      [ 'c', 'z', [0,0,0], [0,9],   ["m","n"] ],
      [ 'm', 'B', [0,0,0], [5,6],   []        ],
      [ 'c', 'u', [0,0,0], [4,1],   ["f"]     ],
      [ 'c', 'v', [0,0,0], [4],     ["g"]     ],
      [ 'm', 'C', [0,0,0], [8,9],   []        ],
      [ 'c', 'j', [0,0,0], [7],     []        ],
      [ 'c', 'k', [0,0,0], [7,3],   []        ]
    ],
    :
  ]
  # List of molecule instances and locations
  [
    [ class, [x,y,z], [nx,ny,nz] ],
    [ class, [x,y,z], [nx,ny,nz] ],
    [ class, [x,y,z], [nx,ny,nz] ],
    [ class, [x,y,z], [nx,ny,nz] ],
    :
    [ class, [x,y,z], [nx,ny,nz] ]
  ]
]

The "class" is just an integer index into the list of molecule
complex templates (0 is first). Since the molecules are already
sorted by species, it might be possible to subdivide the instance
list by the class index (creating one more layer of nesting) and
eliminating the class entry from each instance. However, since
each molecule has a pointer to its own "graph_pattern" it's not
clear that there's any guarantee that the molecules in a group
will all have the same graph pattern. In any event, it will be
relatively easy to subdivide the list by classes in Python.

This format should be importable as either a JSON file or a python
source file for speed testing.